requirements.txt
PyQt5==5.15.4
numpy==1.19.5
pillow==8.1.2
opencv-python==4.5.1.48
scipy==1.5.4
face_recognition==1.3.0
tensorflow==2.4.1
matplotlib=3.3.4
